game is nearly done
cleaned up the code, added functionality to the food, score increase and interactivity with the fox behavior. the food is dragged to the foxes who then eat it, wait to be pet, and if the user clicks on the fox they get score increased again. the foxes then burrow into the ground, either in a pre existing burrow or creating a new one in their current position.

will add next:
- visual representation of gaining or losing points
- score/lives displayed on screen
- high score? instead of refreshing page restart game while keeping track of highest score?
- tutorial sequence???