a few errors but progress
two unknown errors: when I try to use IDs and a scene array to move between scenes, it always messes up, so I went the long way around by manually checking each name for the scene Im switching to and it works now. there must be a more efficient way to do it... but I'm a bit at a loss so I'll ask in class about it.

another error: attempting to drag the item around gives me a bunch of errors which I think are associated to the fact I'm moving the item to the new variable activeItem. Not sure how to fix this while also keeping it effective considering it's pretty helpful for me to have that variable in the first place to be called in other places... this doesn't work like c++ would and its making my head hurt