added sound, deer, fixed statue
statue wouldn't have worked because of the x and y coordinates for the unzoomed statue being what will be passed on to interact with the hand, so I made a new class (PHand) to split the code between both. also fixed coordinates so the hand will display properly when added.

added deer and sound and some more item interactions. deer is fully functional, just missing its item :) unfortunately didn't get to finish radio yet so I'll do that tomorrow

MIGHT move bgm from Start.js to an audio management class if there ends up being a lot of sounds in the game