buttons + radio + fixed inventory
I realized the inventory wasn't going to work out because there's no way to stop adding items if it's full. I made addtoinventory return true or false on whether or not the item was actually added. the bool makes sure items don't get deleted when inventory is full

I also added the radio puzzle drawings (and struggled for a bit to fix the code because the way I wrote it before was just not working out.)

I'll finish room 4 (drawings are almost done) and all the items tomorrow