finished (maybe) (almost)
everything's done!! added window and bird. playtested a few times and realized there was a problem with the door. it took me forever to realize it was just a "this" written where it shouldn't have been. will do a short fade in and out tomorrow for the ending and that's it!